requirements.txt
discord.py
PyNaCl
dotenv
pytube